+++ buweiqi/AI1_Header/PointStyle.h
@@ -7,30 +7,30 @@ class PointStyle
 public:
 	//加分
 	int CurrentFormatPointStyle;
-	int CurrentFormatScorePoor;
+	double CurrentFormatScorePoor;
 	int CurrentPointStyle;
-	int CurrentScorePoor;
+	double CurrentScorePoor;
 	//减分
 	int LastFormatPointStyle;
-	int LastFormatScorePoor;
+	double LastFormatScorePoor;
 	int LastSpecialPointStyle;
-	int LastSpecialScorePoor;
+	double LastSpecialScorePoor;
 	//层差分
-	int RivalSinglePointPoor;//对方单点
-	int RivalChipPoor;//对方缺口
-	int RivalTigerMouthPoor;//对方虎口
+	double RivalSinglePointPoor;//对方单点
+	double RivalChipPoor;//对方缺口
+	double RivalTigerMouthPoor;//对方虎口
 
-	int RivalFormatChipPoor;//对方形成缺口
-	int RivalFormatTigerMouthPoor;//对方形成虎口
-	int RivalFormatEyePoor;//对方形成眼
+	double RivalFormatChipPoor;//对方形成缺口
+	double RivalFormatTigerMouthPoor;//对方形成虎口
+	double RivalFormatEyePoor;//对方形成眼
 
-	int MySinglePointPoor;//己方单点
-	int MyChipPoor;//己方缺口
-	int MyTigerMouthPoor;//己方虎口
+	double MySinglePointPoor;//己方单点
+	double MyChipPoor;//己方缺口
+	double MyTigerMouthPoor;//己方虎口
 
-	int MyFormatChipPoor;//己方形成缺口
-	int MyFormatTigerMouthPoor;//己方形成虎口
-	int MyFormatEyePoor;//己方形成眼
+	double MyFormatChipPoor;//己方形成缺口
+	double MyFormatTigerMouthPoor;//己方形成虎口
+	double MyFormatEyePoor;//己方形成眼
 
 };
